# Use socket option {packet, 4} for stream protocol