The stream protocol frame (packet) starts with a 4 byte
unsigned integer in big-endian byte order.

Therefore we can make use of socket option `{packet, 4}`.
The runtime will generate the header for send operations
and strip off the header for receive operations.
See `inet:setopts/2` docs for more info.

Before this commit, rabbitmq-server received partial
info messages and had to assemble a stream packet on its own.

After this commit, rabbitmq-server receives one info message
per stream frame.

Advantages of this commit:
1. Less copying of binary data for large frames.
Large frames are for example created with large `--size` or
`--batch-size` flags in stream-perf-test.
Hence, this fixes the following TODO:
https://github.com/rabbitmq/rabbitmq-server/blob/744100ed2fcc0b4ce137a9fd9df4ba3ba5839997/deps/rabbitmq_stream_common/src/rabbit_stream_core.erl#L558-L561
2. A good amount of code gets deleted.