@ansd 2022-06-20
Performance tests using `stream-perf-test` with different number for `--size` show that message throughput drops slightly in some cases with this PR.

Running rabbitmq-server with `+S 2` and stream-perf-test for 30 seconds.

Examples:

1.
```
java -jar target/stream-perf-test.jar  -x 1 -y 0 -s 50000 -bs 10 -smssb 3gb -mlb 20gb
```

Before:
```
Summary: published 14825 msg/s, confirmed 14561 msg/s, consumed 0 msg/s, latency 95th 0 ms, chunk size 0
```
After:
```
Summary: published 14391 msg/s, confirmed 14178 msg/s, consumed 0 msg/s, latency 95th 0 ms, chunk size 0
```

2.
```
java -jar target/stream-perf-test.jar -x 1 -y 0 -s 1000 -smssb 3gb -mlb 20gb
```

Before:
```
Summary: published 377109 msg/s, confirmed 376781 msg/s, consumed 0 msg/s, latency 95th 0 ms, chunk size 0
```
After:
```
Summary: published 374069 msg/s, confirmed 373742 msg/s, consumed 0 msg/s, latency 95th 0 ms, chunk size 0
```

I created a minimal example comparing `{packet, raw}` vs `{packet, 4}` without using RabbitMQ at all in https://github.com/ansd/packet-parser which also shows that `{packet, raw}` is slower in some cases.