@ansd 2022-06-29
Closing this.

As shown in https://github.com/ansd/packet-parser and discussed in https://erlangforums.com/t/performance-of-socket-option-packet-4/1562 socket option `{packet, 4}` is too slow for our purposes. We achieve higher throughput by parsing the packets on our own in the application level (instead of in the Erlang runtime).

Instead of using `{packet, 4}`, we should handle the `iodata()` in subsequent calls rather than turning it into a `binary()`.